requirements.txt
asgiref==3.5.2
backcall==0.2.0
beautifulsoup4==4.11.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
decorator==5.1.1
diff-match-patch==20200713
Django==3.1.7
django-celery-beat==2.4.0
django-cors-headers==3.7.0
django-countries==7.4.2
django-extensions==3.1.1
django-filter==2.4.0
django-health-check==3.17.0
django-import-export==2.5.0
django-ses==1.0.3
django-storages==1.11.1
django-timezone-field==5.0
djangorestframework==3.12.2
drf-yasg==1.20.0
grpcio==1.37.0
gunicorn==20.1.0
httplib2==0.21.0
idna==2.9
importlib-metadata==3.7.3
ipython==7.21.0
ipython-genutils==0.2.0
msgpack==1.0.4
numpy==1.23.4
packaging==20.9
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
typing-extensions>=4.0.1
uritemplate==3.0.1
urllib3==1.26.12
wcwidth==0.2.5
xlwt==1.3.0
zipp==3.10.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
whitenoise==6.2.0
openapi-codec==1.3.2
openpyxl==3.0.10